+++ src/utils.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <sys/types.h>
 
-u_int64_t get_symbol_repetition(char c, char *s);
+u_int16_t get_symbol_repetition(char c, char *s);
 int get_file_size(FILE* file);
 
 #endif
